Completed Auto Balancing Command (#3)
* Got base PID configuration and implemented charge station balancing command

Button activation of the command must still be implemented. The command itself must be tested as well.

* Completed Charge Station Balancing Command

The only issue we have to consider is the possibility of another (heavier) robot preventing us from balancing the charge station, which would result in the robot driving over the higher edge of the charge station. The "whileTrue" button press that activates this command means that the driver can stop the command in this case, although we may be able to automate this override in the future.

* Increased robot velocity on ChargeStationBalancingCommand