READ HERE
Basically, the primary issue I have discovered is that In periodic we run the "Go to scoring position method" doing this will result in viewing trajectories in realtime. This was more useful for seeing the accuracy of SwerveOdometry Vs the Limelight botpose readings along with diagnosing false positives. Another thing is we have a button binding to run this method in robotinit. Both of this causes the robot not to drive, IF we remove the button binding and left the method in periodic, the robot would just IMMEDIETLY Drive/Follow trajectory when teleop starts. The tradeoff here is that we may not see the trajectory realtime on glass but our robot would at least work and we start automatically driving/following trajectorys