Changed condition from checking if a substation is selected to detect… (#75)
…ing a game piece

The old condition (if a substation is selected on the shuffleboard)
requires the operator to deselect the substation for the arm to move to
the correct scoring setpoint. The new condition (checking if we have a
game piece) means the robot will know if we intend to score or load, so
the operator does not have to specify.